corrected functionallity for adding multiple audio files and enabled functionality for checking and getting audio files upon collectionItem request

+++ apiHelpers.js
@@ -85,7 +85,9 @@ const googleTextToSpeech = (word, languageCode = 'en') => {
             rej(err);
           }
           if (data) {
-            fs.unlinkSync(`${word}.mp3`);
+            if (fs.exists(`${word}.mp3`)){
+              fs.unlinkSync(`${word}.mp3`);
+            }
             console.log("Uploaded in:", data.Location);
             res(data.Location);
           }
